Project/raspberry/requirements.txt
rpi.gpio
paho-mqtt
keen